server/requirements.txt
python-docx
pymongo
Flask
PyJWT
Werkzeug